delivery_postlogistics: drop transaction rollback/commit
When sending a request for multiple packages, postlogistics might return
an error for one of those.
However, nobody is invoiced while one of the successful is scanned by
swisspost. It means that it's safe if we do not save the successful
labels when one fails.
Raising an exception when a label fails lets odoo manage the rollback.
When user fixed the package, he can then try to confirm the picking
again, and generate new labels.